# backend/openapi-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para la gestión de camisetas deportivas, clientes B2B, tallas y ventas. Permite administrar el stock de camisetas, gestionar clientes preferenciales y regulares, calcular precios con descuentos y registrar ventas.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/tallas": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Listar tallas",
                "description": "Obtiene todas las tallas registradas.",
                "operationId": "listarTallas",
                "responses": {
                    "200": {
                        "description": "Listado de tallas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Talla"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Crear talla",
                "description": "Registra una nueva talla para camisetas.",
                "operationId": "crearTalla",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TallaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talla creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Talla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/tallas/{talla}": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Obtener talla",
                "description": "Obtiene una talla específica.",
                "operationId": "obtenerTalla",
                "parameters": [
                    {
                        "name": "talla",
                        "in": "path",
                        "description": "ID de la talla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Talla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Actualizar talla",
                "description": "Actualiza los datos de una talla existente.",
                "operationId": "actualizarTalla",
                "parameters": [
                    {
                        "name": "talla",
                        "in": "path",
                        "description": "ID de la talla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TallaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Talla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Eliminar talla",
                "description": "Elimina una talla siempre que no esté asociada a camisetas.",
                "operationId": "eliminarTalla",
                "parameters": [
                    {
                        "name": "talla",
                        "in": "path",
                        "description": "ID de la talla",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla eliminada correctamente"
                    },
                    "409": {
                        "description": "La talla está asociada a una o más camisetas"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Talla": {
                "title": "Talla",
                "description": "Modelo completo de talla disponible para camisetas",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "M"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TallaInput": {
                "title": "Talla Input",
                "description": "Datos para crear o actualizar una talla",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "M"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}